Made separate PyronaThread constructor in using.py
